requirements.txt
# Weather RAG LLM - Dependencies
# Python 3.10+

# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1

# HTTP client
httpx==0.26.0

# RAG & embeddings (use OpenAI or local)
langchain==0.1.9
langchain-community==0.0.24
langchain-openai==0.0.6
langchain-text-splitters==0.0.1
langchain-core==0.1.23
chromadb==0.4.22
sentence-transformers==2.3.1

# Environment
python-dotenv==1.0.1

# Optional: for OpenAI embeddings/LLM (set OPENAI_API_KEY in .env)
openai==1.12.0
